refactor: organize security classes in subpackages
