+++ src/main/java/com/pragma/powerup/application/dto/request/UserRoleRequestDto.java
@@ -1,0 +1,12 @@
+package com.pragma.powerup.application.dto.request;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserRoleRequestDto {
+    @Schema(example = "CLIENT")
+    private String name;
+}
