+++ src/main/java/com/pragma/powerup/application/handler/impl/UserHandler.java
@@ -7,7 +7,7 @@ import com.pragma.powerup.application.dto.response.UserResponseDto;
 import com.pragma.powerup.application.handler.IUserHandler;
 import com.pragma.powerup.application.mapper.IUserRequestMapper;
 import com.pragma.powerup.application.mapper.IUserResponseMapper;
-import com.pragma.powerup.infrastructure.security.config.TokenProvider;
+import com.pragma.powerup.infrastructure.security.utils.TokenUtils;
 import com.pragma.powerup.domain.api.IUserServicePort;
 import com.pragma.powerup.domain.model.RoleModel;
 import com.pragma.powerup.domain.model.UserModel;
@@ -32,7 +32,7 @@ public class UserHandler implements IUserHandler {
     private final IUserResponseMapper userResponseMapper;
     private final BCryptPasswordEncoder passwordEncoder;
     private final AuthenticationManager authenticationManager;
-    private final TokenProvider jwtTokenUtil;
+    private final TokenUtils jwtTokenUtil;
 
     @Override
     public AuthTokenResponseDto authenticateUser(UserLoginDto userLoginDto) {
